pyproject.toml
[tool.poetry]
name = "chess"
version = "0.1.0"
description = ""
authors = ["Clément Coutet <clement.coutet@grenoble-inp.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ruff = "^0.7.3"
mypy = "^1.13.0"
pillow = "^11.0.0"
cairosvg = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
